refactor(native): move prompt assembly into its own module
Tool catalog resolution, system prompt assembly, personalization lookup, and
memory-section formatting move to `api_process_adapter/prompt.rs`:
`resolve_tool_catalog_with_code_intelligence`, `resolve_personalization_settings`,
`resolve_system_prompt_with_settings`, `select_memory_bodies`,
`format_system_prompt`, `format_memory_section`, and
`format_custom_instructions_section`.

Measured at 363 lines, not the 440 design.md estimated. That estimate came from
a contiguous line range starting 67 lines earlier, at `turns_character_count` —
four compaction predicates that tasks.md assigns to `compaction.rs`. The census
correction is folded into design.md.

Bodies are byte-identical apart from rustfmt rewrapping
`format_custom_instructions_section`'s signature, which the added `pub(super)`
pushed past 100 columns.
